Use unsigned long instead of DWORD

+++ src/openrct2/util/Util.h
@@ -31,7 +31,7 @@ int32_t MphToDmps(int32_t mph);
 int32_t UtilBitScanForward(uint32_t source)
 {
 #if defined(_MSC_VER) && (_MSC_VER >= 1400) // Visual Studio 2005
-    DWORD i;
+    unsigned long i;
     uint8_t success = _BitScanForward(&i, source);
     return success != 0 ? i : -1;
 #elif defined(__GNUC__)
@@ -53,7 +53,7 @@ int32_t UtilBitScanForward(uint32_t source)
 int32_t UtilBitScanForward(uint64_t source)
 {
 #if defined(_MSC_VER) && (_MSC_VER >= 1400) && defined(_M_X64) // Visual Studio 2005
-    DWORD i;
+    unsigned long i;
     uint8_t success = _BitScanForward64(&i, source);
     return success != 0 ? i : -1;
 #elif defined(__GNUC__)
